+++ package.json
@@ -332,7 +332,6 @@
     "eslint-plugin-promise": "^3.4.0",
     "eslint-plugin-standard": "^3.0.1",
     "expect.js": "^0.3.1",
-    "jsdom": "^9.12.0",
     "karma": "^1.7.1",
     "karma-browserify": "^5.0.5",
     "karma-chrome-launcher": "^2.0.0",
